# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Системы для управления банковскими картами",
    "contact": {
      "name": "Вурц Геннадий",
      "email": "vgd2001@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/admin/card/block": {
      "put": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Блокировка карты",
        "description": "Поддерживается блокировка как по запросу от пользователя, так и без него.\nДля блокировки по запросу пользователя - передать в параметрах Id запроса на блокировку.\nДля блокировки без запроса от пользователя - указать в теле запроса номер карты, в параметрах запроса при этом ничего не указывать.",
        "operationId": "blockCard",
        "parameters": [
          {
            "name": "requestId",
            "in": "query",
            "description": "Номер запроса на блокировку",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Номер карты",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": 1234567876543210
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Карта заблокирована",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDto",
                  "example": {
                    "number": "**** **** **** 1234",
                    "ownerId": 1,
                    "created": "2025-09-28T12:00",
                    "expiration": "2026-09-28",
                    "status": "BLOCKED",
                    "balance": 1234.56
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Необходимо указать id запроса на блокировку либо номер карты"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Не найдена карта или не найден запрос на блокировку",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Не найдена карта": {
                    "description": "Не найдена карта",
                    "value": {
                      "error": "Card with number **** **** **** 1234 not found"
                    }
                  },
                  "Не найден запрос на блокировку": {
                    "description": "Не найден запрос на блокировку",
                    "value": {
                      "error": "Card blocking request with id 42 not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/card/activate": {
      "put": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Активация карты",
        "operationId": "activateCard",
        "requestBody": {
          "description": "Номер карты",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": 1234567876543210
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Карта активирована",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDto"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Card with number **** **** **** 1234 not found"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/card/block": {
      "post": {
        "tags": [
          "User: карты"
        ],
        "summary": "Создать запрос на блокировку карты",
        "operationId": "createBlockingCardRequest",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "description": "Id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Частичный номер карты (например последние 4 цифры)",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": 1234
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Запрос на блокировку карты создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardBlockingRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Card with number 6016 not found"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Ошибка уникальности или целостности данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Запрос на блокировку для этой карты уже существует"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/card": {
      "get": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Поиск карт",
        "description": "Поддерживается фильтрация и пагинация",
        "operationId": "getCards",
        "parameters": [
          {
            "name": "number",
            "in": "query",
            "description": "Номер карты, поддерживается поиск по части номера",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 1234
          },
          {
            "name": "userId",
            "in": "query",
            "description": "ID пользователя",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            },
            "example": 42
          },
          {
            "name": "status",
            "in": "query",
            "description": "Статус карты",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "BLOCKED",
                "EXPIRED"
              ]
            },
            "example": "ACTIVE"
          },
          {
            "name": "created",
            "in": "query",
            "description": "Дата создания карты",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2025-09-28"
          },
          {
            "name": "expiration",
            "in": "query",
            "description": "Дата окончания срока действия карты",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-28"
          },
          {
            "name": "from",
            "in": "query",
            "description": "Индекс первого элемента, с которого нужно вернуть",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество возвращаемых элементов",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список найденных карт",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "from должно быть не меньше 0."
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Создание новой карты",
        "operationId": "createNewCard",
        "requestBody": {
          "description": "Данные для создания новой карты. Поле balance опциональное, expiration должно быть в будущем. После создания карта имеет статус BLOCKED, нужно её активировать.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Карта создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Поле expiration должно содержать дату, которая еще не наступила."
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "User with id 42 not found"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Удаление карты",
        "description": "Запрещено удалять карту имеющую не нулевой баланс",
        "operationId": "deleteCard",
        "requestBody": {
          "description": "Номер карты",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": 1234567876543210
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Карта удалена"
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Card with number **** **** **** 1234 not found"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Ошибка удаления карты",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Card with number **** **** **** 1234 has non-zero balance"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/card/get": {
      "post": {
        "tags": [
          "Админ: Карты"
        ],
        "summary": "Получение информации по карте",
        "description": "Основная информация о карте, включая описание владельца. Использую POST метод для получения данных по карте, для того чтобы не передавать чувствительную информацию в параметрах запроса",
        "operationId": "getCard",
        "requestBody": {
          "description": "Номер карты",
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": 1234567876543210
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Карта найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardFullDto"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "Card with number **** **** **** 1234 not found"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/card": {
      "get": {
        "tags": [
          "User: карты"
        ],
        "summary": "Поиск карт пользователя",
        "description": "Поддерживается фильтрация и пагинация",
        "operationId": "getCards_1",
        "parameters": [
          {
            "name": "number",
            "in": "query",
            "description": "Номер карты, поддерживается поиск по части номера карты",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 1234
          },
          {
            "name": "status",
            "in": "query",
            "description": "Статус карты",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "BLOCKED",
                "EXPIRED"
              ]
            },
            "example": "ACTIVE"
          },
          {
            "name": "created",
            "in": "query",
            "description": "Дата создания карты",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2025-09-28"
          },
          {
            "name": "expiration",
            "in": "query",
            "description": "Дата окончания срока действия карты",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-28"
          },
          {
            "name": "from",
            "in": "query",
            "description": "Индекс первого элемента, с которого нужно вернуть",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество возвращаемых элементов",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "description": "Id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Список найденных карт",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardShortDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "from должно быть не меньше 0."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CardDto": {
        "type": "object",
        "description": "Информация о банковской карте",
        "properties": {
          "number": {
            "type": "string",
            "description": "номер карты",
            "example": "**** **** **** 1234"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64",
            "description": "id владельца карты",
            "example": 1
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "дата создания карты",
            "example": "2025-09-28T12:00"
          },
          "expiration": {
            "type": "string",
            "format": "date",
            "description": "дата окончания срока действия карты",
            "example": "2026-09-28"
          },
          "status": {
            "type": "string",
            "description": "статус карты",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ],
            "example": "ACTIVE"
          },
          "balance": {
            "type": "number",
            "description": "баланс карты",
            "example": 1234.56
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Сведения об ошибке",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки",
            "example": "error message"
          }
        }
      },
...
      "CardBlockingRequestDto": {
        "type": "object",
        "description": "Объект запроса пользователя на блокировку карты",
        "properties": {
          "cardNumber": {
            "type": "string",
            "description": "номер карты",
            "example": "**** **** **** 1234"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "дата и время создания запроса",
            "example": "2025-09-29T12:00"
          },
          "solved": {
            "type": "boolean",
            "description": "статус решения запроса",
            "example": false
          }
        }
      },
...
    },
...
  }
}